SourceMod can now be somewhat compiled on OS X for patch sanity testing (bug 3516, r=ds). This adds the ability for us to change the GCC version we use more flexibly.

+++ extensions/sqlite/driver/SqResults.cpp
@@ -343,7 +343,7 @@ bool SqResults::IsNull(unsigned int columnId)
 	return (field->type == SQLITE_NULL);
 }
 
-unsigned int SqResults::GetDataSize(unsigned int columnId)
+size_t SqResults::GetDataSize(unsigned int columnId)
 {
 	SqField *field = GetField(columnId);
 	if (!field)
